requirements.txt
mattermostdriver==7.3.2
psycopg2-binary==2.9.9
schedule==1.2.0
python-dotenv==1.0.0